requirements.txt
langchain-core
langchain-openai
langchain-together
python-dotenv
tavily-python
yfinance
pandas
matplotlib
plotly
googlenewsdecoder
langgraph
mcp
aucodb
neo4j
pyvis
alpha_vantage
redis
together